chore(validation): move Auth, EnableOTP and OTPAuth validations to auth validation folder
